dynamo: use core e2e tests
